Code change made for BugFix2_deleteError committed

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -149,9 +149,24 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+
+        Image image = imageService.getImage(imageId);
+        User user=(User)session.getAttribute("loggeduser");
+        String error = "Only the owner of the image can delete the image";
+
+        model.addAttribute("image",image);
+        model.addAttribute("deleteError", error);
+        model.addAttribute("user",user);
+
+        if(!user.getId().equals(image.getUser().getId())) {
+            return "images/image";
+        }
+        else{
+
+            imageService.deleteImage(imageId);
+            return "redirect:/images/";
+        }
     }
